fix(keychain): storage access cookie drop use cookie for safari

+++ keychain/src/utils/storage.ts
@@ -1,9 +1,19 @@
+import Cookies from "js-cookie";
+
+const isSafari =
+  typeof navigator !== "undefined" &&
+  /^((?!chrome|android).)*safari/i.test(navigator.userAgent);
+
 const Storage = {
   keys: (): string[] => {
-    return Object.keys(window.localStorage);
+    if (!isSafari && typeof window != "undefined") {
+      return Object.keys(window.localStorage);
+    }
+
+    return Object.keys(Cookies.get());
   },
   get: (key: string): any => {
-    if (typeof window != "undefined") {
+    if (!isSafari && typeof window != "undefined") {
       const value = window.localStorage.getItem(key);
       if (!value) {
         return null;
@@ -11,12 +21,25 @@ const Storage = {
 
       return JSON.parse(value);
     }
+
+    const existing = Cookies.get(key);
+
+    if (typeof existing === "undefined") {
+      return undefined;
+    }
+
+    return JSON.parse(existing);
   },
   set: (key: string, value: any) => {
-    if (typeof window != "undefined") {
+    if (!isSafari && typeof window != "undefined") {
       window.localStorage.setItem(key, JSON.stringify(value));
       return;
     }
+
+    Cookies.set(key, JSON.stringify(value), {
+      secure: true,
+      sameSite: "strict",
+    });
   },
   update: (key: string, value: any) => {
     const existing = Storage.get(key);
@@ -26,16 +49,21 @@ const Storage = {
     Storage.set(key, value);
   },
   remove: (key: string) => {
-    if (typeof window != "undefined") {
+    if (!isSafari && typeof window != "undefined") {
       window.localStorage.removeItem(key);
       return;
     }
+
+    Cookies.remove(key);
   },
   clear: () => {
-    if (typeof window != "undefined") {
+    if (!isSafari && typeof window != "undefined") {
       window.localStorage.clear();
       return;
     }
+
+    const cookies = Cookies.get();
+    Object.keys(cookies).forEach((key) => Cookies.remove(key));
   },
 };
